moved code into the spritebot module
