+++ .flake8
@@ -1,0 +1,2 @@
+[flake8]
+ignore = E501, F841, E303, E226, E128, E722, E265, W504, E302, E126, E302, E305
