--- Constants.py
@@ -1,31 +1,0 @@
-
-PORTRAIT_SIZE = 0
-PORTRAIT_TILE_X = 0
-PORTRAIT_TILE_Y = 0
-
-PORTRAIT_SHEET_WIDTH = PORTRAIT_SIZE * PORTRAIT_TILE_X
-PORTRAIT_SHEET_HEIGHT = PORTRAIT_SIZE * PORTRAIT_TILE_Y
-
-CROP_PORTRAITS = True
-
-COMPLETION_EMOTIONS = []
-
-EMOTIONS = []
-
-
-ACTION_MAP = { }
-
-COMPLETION_ACTIONS = []
-
-ACTIONS = []
-
-DIRECTIONS = [ "Down",
-               "DownRight",
-               "Right",
-               "UpRight",
-               "Up",
-               "UpLeft",
-               "Left",
-               "DownLeft"]
-
-MULTI_SHEET_XML = "AnimData.xml"
