+++ spritebot/utils.py
